% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_farmslist.R
\name{create_farmslist}
\alias{create_farmslist}
\title{Create farm list}
\usage{
create_farmslist(listfarms, fi2ges, fa2pw, pw2fi)
}
\arguments{
\item{qIng}{Average quantity of faeces ingested by a pig (g/day)}

\item{dataframefarms}{Data.frame listing farms and their
characteristics: id, type, nbRP, nbPW, nbFat}

\item{fa2pw}{data.frame with three columns: tail, head and p.  Each
row represent a possible connection from the tail to the head
having a probability p to happen in case of external fa-pw
movement.}

\item{pw2fi}{data.frame with three columns: tail, head and p.  Each
row represent a possible connection from the tail to the head
having a probability p to happen in case of external pw-fi
movement.}

\item{fi2ges_w2}{data.frame with three columns: tail, head and p.
Each row represent a possible connection from the tail to the
head having a probability p to happen in case of external
fi-ges movement.}

\item{beta_ges}{Transmission rate in the gestation sector (default
value = 0.7),}

\item{beta_fa}{Transmission rate in the farrowing sector (default
value = 0.7),}

\item{beta_pw}{Transmission rate in the post-weaning sector
(default value = 0.7),}

\item{beta_fi}{Transmission rate in the finishing sector (default
value = 0.7)}
}
\value{
List
}
\description{
\code{create_farmslist} Generate the list of farms
    with export destination characteristics
}
\examples{
farmslist <- create_farmslist(listfarms, fi2ges, fa2pw, pw2fi)

}
