% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{HEV}
\alias{HEV}
\title{Create a model for the SimInf framework}
\usage{
HEV(
  farms,
  tspan = NULL,
  days = NULL,
  lambda = 1/45.8,
  decRate = 0.08,
  SFRate = 0.7,
  qIng = 25,
  mu = 1/30,
  qFeces_sow = 2000,
  qFeces_pig = 550
)
}
\arguments{
\item{farms}{List of farms.}

\item{tspan}{A vector of increasing time points
where the state of each node is to be returned.}

\item{days}{the number of days to generate events.}
}
\description{
Create a model to be used by the SimInf framework.
}
