% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_nodes.R
\name{network_nodes}
\alias{network_nodes}
\title{Sum shipped animals}
\usage{
network_nodes(data)
}
\arguments{
\item{dataset}{data.frame including transported animal number and
origin and destination ids, sectors and farming types columns}
}
\value{
data.frame
}
\description{
\code{network_nodes} sum animals shipped from a given
    sector of a given farm to a given sector of a given farm
}
\examples{
result_nAnim <- network_nodes(data)

}
