% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplecoinffarms.R
\name{samplecoinffarms}
\alias{samplecoinffarms}
\title{Sample co-infected farms}
\usage{
samplecoinffarms(farmlist, free_prob, id, farmtype, sampledtyp)
}
\arguments{
\item{farmlist}{data.frame including type and id columns}

\item{free_prob}{probability of FF farms to be IMV-free}

\item{id}{column name indicating site id in the farmlist dataframe}

\item{farmtype}{column name indicating farm type in the farmlist dataframe}

\item{sampledtype}{type of farms amoung with coinfected farms will be sampled}
}
\value{
Vector containing the sampled farms id
}
\description{
\code{samplecoinffarms} Create a subset of farms that are co-infected
}
\examples{
free_FF <- samplecoinffarms(farmlist, free_prob)

}
