% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slaughter_info.R
\name{slaughter_info}
\alias{slaughter_info}
\title{Get the time points when slaughter happens}
\usage{
slaughter_info(events, day_intro)
}
\arguments{
\item{events}{data.frame of events}

\item{day_intro}{day of exposed animal introduction}
}
\value{
Two numbers vector
}
\description{
\code{slaughter_info}  Get the time points when slaughter happens
}
\examples{
tspan <- slaughter_info(events, day_intro)

}
