% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyBRS.R
\name{specifyBRS}
\alias{specifyBRS}
\title{Characterize the batch rearing systems}
\usage{
FIX ME
}
\arguments{
\item{id.BRS}{FIX ME}

\item{nbatches}{FIX ME}

\item{size_ges_sector}{FIX ME}

\item{size_fa_sector}{FIX ME}

\item{size_pw_sector}{FIX ME}

\item{size_fi_sector}{FIX ME}

\item{duration_ges}{FIX ME}

\item{duration_pre_birth}{FIX ME}

\item{duration_fa}{FIX ME}

\item{duration_pw}{FIX ME}

\item{duration_fi}{FIX ME}

\item{duration_post_ges}{FIX ME}

\item{duration_post_fa}{FIX ME}

\item{duration_post_pw}{FIX ME}

\item{duration_post_fi}{FIX ME}

\item{batches.intervals}{FIX ME}
}
\value{
FIX ME
}
\description{
\code{specifyBRS} FIX ME
}
\examples{
BRS.str <- specifyBRS()

}
